TianYang:handle a possible error
git-svn-id: https://source.kii.com/servo/platform/sdk-mobile-client/androidV1/Samples/KiiFileStorageSampleApp@22420 e7cb0379-d053-0e4b-a1ac-43c7dbffadbe

+++ src/com/kii/cloud/engine/CloudCallback.java
@@ -67,7 +67,9 @@ public class CloudCallback extends KiiFileCallBack {
                 @Override
                 public void run() {
                     try {
-                        f.delete();
+                        if (f != null) {
+                            f.delete();
+                        }
                     } catch (Exception e) {
                         e.printStackTrace();
                     }
